move App.tsx into src directory
