+++ src/App.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { SafeAreaView, StatusBar, useColorScheme } from "react-native";
+
+import { Colors } from "react-native/Libraries/NewAppScreen";
+import { BasePage } from "./page";
+import { flexChild } from "./styles";
+
+function App(): JSX.Element {
+  const isDarkMode = useColorScheme() === "dark";
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+
+  return (
+    <SafeAreaView style={flexChild}>
+      <StatusBar barStyle={isDarkMode ? "light-content" : "dark-content"} backgroundColor={backgroundStyle.backgroundColor} />
+      <BasePage />
+    </SafeAreaView>
+  );
+}
+
+export default App;
